+++ CHANGELOG.md
@@ -1,7 +1,23 @@
 # Change Log
 
 
-## [0.0.6] - unreleased yet
+## [0.0.7] - unreleased yet
+
+
+## [0.0.7] - 2018-04-23
+### Changed
+ * update requirements
+
+### Added
+ * [DOCS]: installation-linker.md
+
+
+## [0.0.6] - 2018-01-26
+
+### Changed
+ * [HASHER] - improved error checking
+ * [HASHER] - removed hack for PyInstaller
+
 
 ## [0.0.5] - 2017-09-18